格式化代码 添加选举部分实现

+++ src/main/java/com/wbq/raft/Log.java
@@ -19,5 +19,5 @@ public interface Log {
 
     LogEntry lastLog();
 
-    int lastLogIndex();
+    long lastLogIndex();
 }
